Split clusterhealth.py into one-endpoint-per-file
Split clusterhealth.py into clusterhealth_config.py and
clusterhealth_status.py to enforce one endpoint class per file.
Update test imports to reference the new module paths.
